Поставки

+++ Shop/WebShop.aspx.cs
@@ -8,12 +8,13 @@ namespace Shop
 {
     public partial class WebForm1 : System.Web.UI.Page
     {
-        string connectionString = ConfigurationManager.ConnectionStrings["connectionString"].ConnectionString;
+        string connectionString;
+
         protected void Page_Load(object sender, EventArgs e)
         {
-
+            connectionString = ConfigurationManager.ConnectionStrings["connectionString"].ConnectionString;
         }
-
+       
         protected void TextBox1_TextChanged(object sender, EventArgs e)
         {
 
@@ -38,19 +39,17 @@ namespace Shop
 
         protected void Delete_Click(object sender, EventArgs e)
         {
-            string productId = GridView1.SelectedDataKey["Id"].ToString();
-            string tableName = "Products";
             var service = new BalanceProductService(connectionString);
+            if (GridView1.SelectedDataKey == null)
+            {
+                return;
+            }
+            string productId = GridView1.SelectedDataKey["Id"].ToString();
+               string tableName = "Products";
             service.DeleteFromTable(tableName, productId);
             GridView1.DataBind();
         }
-
         protected void GridView1_SelectedIndexChanged(object sender, EventArgs e)
-        {
-
-        }
-
-        protected void SqlDataSource1_Selecting(object sender, System.Web.UI.WebControls.SqlDataSourceSelectingEventArgs e)
         {
 
         }
